Explicitly define flags property
By explicitly defining the `flags` class property and setting it to an empty array we can prevent errors when not using the flags feature.

Prior to this change the lack of using flags (i.e. not setting any flags) would cause an error since the Client code ALWAYS calls `$this->setFlags`, which in turn always calls `$query->getFlags()`.